requirements.txt
# Core Agent Framework
crewai>=0.80.0
crewai-tools>=0.14.0

# Data Sources
yfinance>=0.2.36

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# RAG (Retrieval-Augmented Generation)
chromadb>=0.4.0

# Web UI
streamlit>=1.30.0
plotly>=5.18.0

# Testing
pytest>=7.0.0

# Evaluation (also serves as TF-IDF fallback for RAG)
scikit-learn>=1.3.0
